refactor: clean up code and reorganize folder structure
