+++ components/voxFlow/AuthModal.tsx
@@ -3,7 +3,7 @@
 import { useState } from 'react'
 import { Button } from '@/components/ui/button'
 import { useToast } from '@/hooks/use-toast'
-import { useVoxLogin, useVoxRegister } from '@/hooks/useVoxFlow'
+import { useVoxLogin, useVoxRegister } from '@/hooks/use-vox-flow'
 import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle } from '@/components/ui/dialog'
 import { Label } from '@/components/ui/label'
 import { Input } from '@/components/ui/input'
