+++ components/voxFlow/TTSTab.tsx
@@ -3,7 +3,7 @@
 import { useState } from 'react'
 import { Button } from '@/components/ui/button'
 import { useToast } from '@/hooks/use-toast'
-import { useVoxTTS } from '@/hooks/useVoxFlow'
+import { useVoxTTS } from '@/hooks/use-vox-flow'
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
 import { Label } from '@/components/ui/label'
 import { Textarea } from '@/components/ui/textarea'
