+++ components/voxFlow/TranslateTab.tsx
@@ -3,7 +3,7 @@
 import { useState } from 'react'
 import { Button } from '@/components/ui/button'
 import { useToast } from '@/hooks/use-toast'
-import { useVoxTranslate } from '@/hooks/useVoxFlow'
+import { useVoxTranslate } from '@/hooks/use-vox-flow'
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card'
 import { Label } from '@/components/ui/label'
 import { Textarea } from '@/components/ui/textarea'
